Fix CI on Windows
First, the package does not support symlinks on Windows. The files involving symlinks are moved to new directory `tests/files-symlinks` and that is excluded from the test paths (`testPaths`) for `tar/untar/tar` on Windows.

Second, for some reason, on Windows, if the test paths for `tar/untar/tar` include `.stack-work`, a 'permission denied' IO error occurs. So `.stack-work` is excluded from the test paths on Windows.